requirements.txt
flask
flask-cors
python-dotenv
rich